Let primitives have error handling, and more primitives & forms
The thing holding me back was that I would need the modules Error and
Values to depend on each other (since a primitive is a value, and can
throw an error, and errors refer to values ..), so I moved the errors
into the Values module to be done with it.

I quite like the definition of boolBinOp. It makes me happy.

Also: added (if ..) and (begin ..)